requirements.txt
Django~=5.1.1
python-dotenv~=1.0.1
python-ulid-django
djangorestframework~=3.15.2
aiogram~=3.13.0
httpx~=0.27.2
django-filter~=24.3
aiogram_dialog~=2.2.0
requests~=2.32.3
djangorestframework-simplejwt
drf-spectacular